removed all ui libraries, I'm going to use imgui with sdl_renderer backend. easiest method

+++ emulator_main.h
@@ -6,7 +6,7 @@ typedef struct {
 	// HANDLE mem_pipe_handle;
 	int argc;
 	char** argv;
-	u16* breakpoint_arr;
+	int* breakpoint_arr;
 
 } args;
